chore: move export mappers to its usage
since these mappers purpose is to define what the exported payloads look like,
they should live with the controllers using them
